+++ src/main/java/service/ConversationService.java
@@ -11,35 +11,35 @@ public class ConversationService {
 	
 	//3. Quan li cuoc hoi thoai
 	// Lấy conversation theo id
-    public Conversation getConversationById(Long conversationId) {
+    public Conversation getConversationById(Integer conversationId) {
         return conversationDao.getConversation(conversationId);
     }
     
-    public List<Conversation> listConversationsByUser(Long userId) {
+    public List<Conversation> listConversationsByUser(Integer userId) {
         return conversationDao.listByUser(userId);
     }
     
  // Kiểm tra đã có direct conversation giữa 2 user chưa
-    public Conversation getDirectConversation(Long userAId, Long userBId) {
+    public Conversation getDirectConversation(Integer userAId, Integer userBId) {
         return conversationDao.getDirectConversation(userAId, userBId);
     }
     
  // Tạo direct conversation mới (nếu chưa có)
-    public Conversation createDirectConversation(Long userAId, Long userBId) {
+    public Conversation createDirectConversation(Integer userAId, Integer userBId) {
         return conversationDao.createDirectConversation(userAId, userBId);
     }
     
  // Lấy danh sách participants trong conversation
-    public List<Users> listParticipants(Long conversationId) {
+    public List<Users> listParticipants(Integer conversationId) {
         return conversationDao.listParticipants(conversationId);
     }
     
     // Cập nhật tên nhóm (chỉ group conversation)
-    public boolean updateConversationName(Long conversationId, String newName, Long requestUserId) {
+    public boolean updateConversationName(Integer conversationId, String newName, Integer requestUserId) {
         return conversationDao.updateConversation(conversationId, newName, requestUserId);
     }
     
-    // public boolean deleteConversation(Long conversationId, Long requestUserId) {
+    // public boolean deleteConversation(Integer conversationId, Integer requestUserId) {
     //     return conversationDao.deleteConversation(conversationId, requestUserId);
     // }
